# public/tools/schema/openapi_rest_demo.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST demo",
    "description": "This is a sample REST serverform jaktestowac.pl .",
    "termsOfService": "http://jaktestowac.pl/",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Get articles",
        "operationId": "getArticles",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "date"
      },
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Add article",
        "operationId": "postArticle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "422": {
            "description": "Invalid article supplied",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "date"
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get users",
        "operationId": "getUsers",
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "operationId": "postUser",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "409": {
            "description": "Conflict - Invalid user email",
            "content": {}
          },
          "422": {
            "description": "Invalid user supplied",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "Get comments",
        "operationId": "getComments",
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Create comment",
        "operationId": "createComment",
        "requestBody": {
          "description": "Created Comment object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - malformed JSON",
            "content": {}
          },
          "422": {
            "description": "Invalid comment supplied",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "body": {
            "type": "string",
            "description": "Article body"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "image": {
            "type": "string",
            "description": "Image path"
          }
        },
        "xml": {
          "name": "Article"
        }
      },
      "ArticleUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title"
          },
          "body": {
            "type": "string",
            "description": "Article body"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "image": {
            "type": "string",
            "description": "Image path"
          }
        },
        "xml": {
          "name": "Article"
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "article_id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string",
            "description": "Comment body"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        },
        "xml": {
          "name": "Comment"
        }
      },
      "CommentUpdate": {
        "type": "object",
        "properties": {
          "article_id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string",
            "description": "Comment body"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        },
        "xml": {
          "name": "Comment"
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      }
    }
  }
}